requirements.txt
WTForms~=3.2.1
Werkzeug~=3.1.3
Flask~=3.1.0
openpyxl~=3.1.5
SQLAlchemy~=2.0.38
matplotlib~=3.10.0
python-dateutil~=2.9.0.post0
babel~=2.17.0
Flask-Migrate~=4.1.0
alembic~=1.14.1
python-dotenv~=1.0.0
Flask-SQLAlchemy~=3.0.0